% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jotto-package.R
\docType{package}
\name{jotto-package}
\alias{jotto}
\alias{jotto-package}
\title{jotto: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
